# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToNote App OpenApi API Documentation",
        "description": "ToNote App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://getwordprime.herokuapp.com/",
            "description": "Staging API server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/bonuses": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all bonuses",
                "description": "Get all bonuses",
                "operationId": "allBonuss",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post new bonuses",
                "description": "Post new bonuses",
                "operationId": "postBonuss",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BonusStoreFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bonuses/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Show an bonus",
                "description": "Show an bonus",
                "operationId": "showBonus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bonus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update an bonus",
                "description": "Update an bonus",
                "operationId": "BonusUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "bonus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BonusUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete an bonus",
                "description": "Delete an bonus",
                "operationId": "deleteBonus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bonus ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BonusStoreFormRequest": {
                "title": "Bonus Store Form Request Fields",
                "description": "Bonus Store request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "example": "name"
                    },
                    "description": {
                        "title": "description",
                        "description": "description ",
                        "type": "string",
                        "example": "description"
                    },
                    "url": {
                        "title": "url",
                        "description": "url",
                        "type": "string",
                        "example": "https://"
                    }
                },
                "type": "object"
            },
            "BonusUpdateFormRequest": {
                "title": "Bonus Update Form Request Fields",
                "description": "Bonus Update request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "example": "name"
                    },
                    "description": {
                        "title": "description",
                        "description": "description ",
                        "type": "string",
                        "example": "description"
                    },
                    "url": {
                        "title": "url",
                        "description": "url",
                        "type": "string",
                        "example": "https://"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}